+++ sdk/metric/aggregator/lastvalue/lastvalue_test.go
@@ -28,7 +28,7 @@ import (
 	ottest "go.opentelemetry.io/otel/internal/testing"
 	export "go.opentelemetry.io/otel/sdk/export/metric"
 	"go.opentelemetry.io/otel/sdk/export/metric/aggregation"
-	"go.opentelemetry.io/otel/sdk/metric/aggregator/test"
+	"go.opentelemetry.io/otel/sdk/metric/aggregator/aggregatortest"
 )
 
 const count = 100
@@ -68,16 +68,16 @@ func checkZero(t *testing.T, agg *Aggregator) {
 }
 
 func TestLastValueUpdate(t *testing.T) {
-	test.RunProfiles(t, func(t *testing.T, profile test.Profile) {
+	aggregatortest.RunProfiles(t, func(t *testing.T, profile aggregatortest.Profile) {
 		agg, ckpt := new2()
 
-		record := test.NewAggregatorTest(metric.ValueObserverKind, profile.NumberKind)
+		record := aggregatortest.NewAggregatorTest(metric.ValueObserverKind, profile.NumberKind)
 
 		var last metric.Number
 		for i := 0; i < count; i++ {
 			x := profile.Random(rand.Intn(1)*2 - 1)
 			last = x
-			test.CheckedUpdate(t, agg, x, record)
+			aggregatortest.CheckedUpdate(t, agg, x, record)
 		}
 
 		err := agg.SynchronizedMove(ckpt, record)
@@ -90,17 +90,17 @@ func TestLastValueUpdate(t *testing.T) {
 }
 
 func TestLastValueMerge(t *testing.T) {
-	test.RunProfiles(t, func(t *testing.T, profile test.Profile) {
+	aggregatortest.RunProfiles(t, func(t *testing.T, profile aggregatortest.Profile) {
 		agg1, agg2, ckpt1, ckpt2 := new4()
 
-		descriptor := test.NewAggregatorTest(metric.ValueObserverKind, profile.NumberKind)
+		descriptor := aggregatortest.NewAggregatorTest(metric.ValueObserverKind, profile.NumberKind)
 
 		first1 := profile.Random(+1)
 		first2 := profile.Random(+1)
 		first1.AddNumber(profile.NumberKind, first2)
 
-		test.CheckedUpdate(t, agg1, first1, descriptor)
-		test.CheckedUpdate(t, agg2, first2, descriptor)
+		aggregatortest.CheckedUpdate(t, agg1, first1, descriptor)
+		aggregatortest.CheckedUpdate(t, agg2, first2, descriptor)
 
 		require.NoError(t, agg1.SynchronizedMove(ckpt1, descriptor))
 		require.NoError(t, agg2.SynchronizedMove(ckpt2, descriptor))
@@ -114,7 +114,7 @@ func TestLastValueMerge(t *testing.T) {
 		require.Nil(t, err)
 		require.True(t, t1.Before(t2))
 
-		test.CheckedMerge(t, ckpt1, ckpt2, descriptor)
+		aggregatortest.CheckedMerge(t, ckpt1, ckpt2, descriptor)
 
 		lv, ts, err := ckpt1.LastValue()
 		require.Nil(t, err)
@@ -124,7 +124,7 @@ func TestLastValueMerge(t *testing.T) {
 }
 
 func TestLastValueNotSet(t *testing.T) {
-	descriptor := test.NewAggregatorTest(metric.ValueObserverKind, metric.Int64NumberKind)
+	descriptor := aggregatortest.NewAggregatorTest(metric.ValueObserverKind, metric.Int64NumberKind)
 
 	g, ckpt := new2()
 	require.NoError(t, g.SynchronizedMove(ckpt, descriptor))